Revert ":window: :tada: show FCP enrollment banner on create connection page" (#4834)
