+++ airbyte-webapp/src/packages/cloud/components/experiments/FreeConnectorProgram/InlineEnrollmentCallout.module.scss
@@ -6,10 +6,6 @@
   width: 100%;
 }
 
-.withBottomMargin {
-  margin-bottom: variables.$spacing-lg;
-}
-
 .enrollLink {
   text-decoration: underline;
   cursor: pointer;
